Updated license text to BSD 3-clause.
The code written by me was originally published under GPL Version 2,
then GPL Version 3, but not all files were updated to reflect that change.
So the source text was a mix of both.

I searched all of the source code and updated everything to the BSD license.
The exceptions are the items I copied from UCBLogo which are the help text and
the Logo standard library. Those are released under GPL Version 3.