MainWindow can now receive longer messages.
Interprocess communication has a sort-of upper bound to the size of
data that can be received at a time. Some QLogo messages (the ones
that include images) can become bigger than the system limit.
The system compensates by breaking a large block into a series of
smaller blocks. A QProcess will send a signal for each block.
QLogo-GUI will now wait for a complete message before processing it.