Replaced sendmessage() function with message struct.
The method for sending a message from one process to the other was by calling
a function with a lambda parameter. The lambda function would take a QDataStream
object and stream the message objects to it. The function operated by setting up
the QDataStream, calling the lambda with the QDataStream, and then senting the
binary result of the stream operations.

The new way is to create a message object that allows QDataStream operations to
it. Then the object destructor will send the binary to the other process. This
results in easier to follow code.