Fixed bug that prevented using library procedures while running a procedure.
Previously, the Logo Standard Library was loaded at the beginning of execution.

Currently, procedures from the library are loaded on an as-needed basis. A
library procedure is loaded by executing the relevant text which contains
either a "TO" or ".MACRO" statement. The complication is that procedures
cannot be defined while a procedure is running. This change introduces a
temporary and empty procedure frame for the duration of the library procedure
definition.